Make textarea rotating work as expected with recent Pillow
Since Pillow 3.0.0, Image.paste() causes "images do not match"
failure, if size of "mask" does not match against pasted image.

This is root cause that blockdiag does not work as expected with
recent Pillow, if "rotate" attribute of a node is 90, 270, or so.

On the other hand, Image.rotate(expand=True) of Pillow earlier than
3.3.0 (including 2.x, in fact) returns image object with unexpected
size: for example, rotating 10x20 by 90 causes not 20x10 but 21x11.

Therefore, this commit does below for rotating textarea, in order to
catch up recent Pillow.

  - specify expand=True at Image.rotate()
  - crop rotated mask image, if it does not match against pasted image