Use word wrapping rather than character wrapping
Currently, the output diagrams will wrap words at a character boundary
rather than a word boundary. This results in some rather ugly diagrams.
Change this.

Note that this patch was previously submitted against the BitBucket
repo [1]. That went unmerged so we've been carrying a version of this in
OpenStack Nova for a few years [2]. It's pretty battle tested.

[1] https://bitbucket.org/blockdiag/blockdiag/pull-requests/16/
[2] https://github.com/openstack/nova/blob/20.0.0/doc/source/conf.py#L170-L233

Signed-off-by: Stephen Finucane <stephen@that.guru>