remove replacment of codecs.getreader on py
The blockdiag.utils.compat module currently replaces the
defintion of codecs.getreader on python3 with a lamda.
The lamda incorrectly tries to dereference a buffer atirrbute
on the past in io stream that does not exists.

This change remove the replacment of the stardard lib function
which is both incorrect and breaks user of blockdiag that expect
the standard defintion.