tests: Use TrueType font for unicode tests
As noted at [1], the default built-in bitmap font in PIL is latin-1
only. This can result in build failures like the below.

  UnicodeEncodeError: 'latin-1' codec can't encode character '\u3042' in
  position 0: ordinal not in range(256)

I imagine PIL is built differently on different platforms since we don't
see this issue in Travis, but it can be seen on Fedora 31 at least.
Fortunately, we already have a TrueType font in-tree, we can simply
start using this, meaning this test will not attempt to use the bitmap
font.

[1] https://github.com/python-pillow/Pillow/issues/2779#issuecomment-333798938

Signed-off-by: Stephen Finucane <stephen@that.guru>